include cell to ECM alignment
I made some minor adaptations to the set-up of the code (no export, no parameter scan). Also included that the cells direction of motion is the average of the underlying ECM field. Tried the feedback mechanism but that does not work yet. We might need to think on how this feedback works.